Refactor classes to own packages
